+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 - Add aspect ratio to statistics
 - Fix orientation to keep all values in stats table even when filtered
+- Upgrade to react-scripts 5
 
 ## [0.5.0] - 2021-12-23
 - Add version number (v1) to API path